refactor(session): use dynamic import for subcommand
