+++ packages/opencode/src/index.ts
@@ -28,7 +28,7 @@ import { AcpCommand } from "./cli/cmd/acp/command"
 import { EOL } from "os"
 import { WebCommand } from "./cli/cmd/web"
 import { PrCommand } from "./cli/cmd/pr/command"
-import { SessionCommand } from "./cli/cmd/session"
+import { SessionCommand } from "./cli/cmd/session/command"
 import { DbCommand } from "./cli/cmd/db/command"
 import path from "path"
 import { Global } from "@opencode-ai/core/global"
